+++ GUI.cpp
@@ -16,6 +16,18 @@ void Color(int y) {
     SetConsoleTextAttribute(out, y);
 }
 
+void noCursor()
+{
+    CONSOLE_CURSOR_INFO info;
+    HANDLE  out;
+
+    info.bVisible = 0;
+    info.dwSize = 1;
+
+    out = GetStdHandle(STD_OUTPUT_HANDLE);
+    SetConsoleCursorInfo(out, &info);
+}
+
  void simple_square(int X, int Y, int width, int height){
   for (int set1 = 0; set1<height; set1++)
   {
@@ -47,16 +59,6 @@ void Color(int y) {
          Color(7);
      }
  }
-
- void L_block(int X, int Y) {
-     go(X,Y);
-     Color(10);
-     cout << char(220) << char(220) << char(220) << char(220) << endl;
-     go(X, Y+1);
-     Color(10);
-     cout << char(223) << endl;
- }
-
 
  //Quellen:
  // https://mycodecollection.blogspot.com/2015/01/c-console-basic-graphic-techniques.html
