+++ GUI.h
@@ -1,7 +1,14 @@
 #pragma once
+
+
+/// <summary>
+/// Diese Funktion sorgt dafür, dass der Cursor ausgeblendet wird
+/// </summary>
+void noCursor();
+
+
 //Funktion definieren
 void simple_square(int X, int Y, int width, int height);
 void U_square(int X, int Y, int width, int height);
 void go(int a, int b);
 void Color(int y);
-void block(int X, int Y);
