--- Hilfsfunktionen.h
@@ -1,6 +1,0 @@
-#pragma once
-
-/// <summary>
-/// Diese Funktion sorgt dafür, dass der Cursor ausgeblendet wird
-/// </summary>
-void noCursor();
