+++ Praktikum_KW2_Spiel.vcxproj
@@ -128,14 +128,12 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="GUI.cpp" />
-    <ClCompile Include="Hilfsfunktionen.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="Bausteine.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Bausteine.h" />
     <ClInclude Include="GUI.h" />
-    <ClInclude Include="Hilfsfunktionen.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
