+++ Praktikum_KW2_Spiel.vcxproj.filters
@@ -21,18 +21,12 @@
     <ClCompile Include="Bausteine.cpp">
       <Filter>Quelldateien</Filter>
     </ClCompile>
-    <ClCompile Include="Hilfsfunktionen.cpp">
-      <Filter>Quelldateien</Filter>
-    </ClCompile>
     <ClCompile Include="GUI.cpp">
       <Filter>Quelldateien</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Bausteine.h">
-      <Filter>Headerdateien</Filter>
-    </ClInclude>
-    <ClInclude Include="Hilfsfunktionen.h">
       <Filter>Headerdateien</Filter>
     </ClInclude>
     <ClInclude Include="GUI.h">
